backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
torch>=2.2.0
torchvision>=0.17.0
opencv-python-headless>=4.9.0.80

Pillow>=10.2.0
numpy>=1.26.4
scikit-learn>=1.4.0
joblib>=1.3.2
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.27
python-dotenv>=1.0.1

pydantic>=2.6.1
pydantic-settings>=2.1.0
aiofiles>=23.2.1
